requirements.txt
httpx>=0.27,<1
atproto>=0.0.50,<1
cryptography>=43.0,<45
cbor2>=5.6,<6
python-igraph>=0.11; extra == "graph"